refactor: isolate gigachat transport compatibility hooks
